2.0.0.2:  frame_processor_device:   - No longer blocks create_frame if output queue is filled.
git-svn-id: https://casparcg.svn.sourceforge.net/svnroot/casparcg/server/branches/2.0.0.2@246 362d55ac-95cf-4e76-9f9a-cbaa9c17b72d

+++ core/format/pixel_format.h
@@ -43,9 +43,9 @@ struct pixel_format_desc
 	pixel_format::type pix_fmt;
 	std::array<plane, 4> planes;
 
-	static size_t hash(const pixel_format_desc& desc)
+	static int hash(const pixel_format_desc& desc)
 	{
-		size_t hash = 0;
+		int hash = 0;
 		switch(desc.pix_fmt)
 		{
 		case pixel_format::ycbcr:
@@ -57,10 +57,10 @@ struct pixel_format_desc
 		    // 28-29 (2)  unused
 		    // 30	 (1)  alpha
 		    // 31    (1)  yuv = true => 1
-			hash |= (desc.planes[0].width & 0x7FF) << 0;
-			hash |= (desc.planes[0].height & 0x7FF) << 11;
-			hash |= ((desc.planes[0].height/desc.planes[1].height) & 0x7) << 22;
-			hash |= ((desc.planes[0].width/desc.planes[1].width) & 0x7) << 25;
+			hash |= ( desc.planes[0].width							& 0x7FF	) << 0;
+			hash |= ( desc.planes[0].height							& 0x7FF	) << 11;
+			hash |= ((desc.planes[0].height/desc.planes[1].height)	& 0x7	) << 22;
+			hash |= ((desc.planes[0].width/desc.planes[1].width)	& 0x7	) << 25;
 			hash |= desc.pix_fmt == pixel_format::ycbcra ? (1 << 30) : 0;
 			hash |= 1 << 31;
 			return hash;
